Rename verusPayInvoiceDetailsHandler and add verus:// to android manifest and ios info plist
