+++ ios/verusMobile/Info.plist
@@ -26,6 +26,7 @@
 			<key>CFBundleURLSchemes</key>
 			<array>
 				<string>i5jtwbp6zymeay9llnraglgjqgdrffsau4</string>
+        <string>verus</string>
 			</array>
 		</dict>
 	</array>
